saving the new structure
